Commit tutorial django part 6 and move test in view to test.py and fig an error
