+++ polls/static/polls/style.css
@@ -1,0 +1,7 @@
+li a {
+    color: green;
+}
+
+body {
+    background: white url("images/background.gif") no-repeat;
+}
